+++ firmware/application/src/app_cmd.h
@@ -2,7 +2,9 @@
 #define APP_CMD_H
 
 #include <stdint.h>
+#include <stdbool.h>
 #include "dataframe.h"
+#include "settings.h"
 
 
 typedef data_frame_tx_t *(*cmd_processor)(uint16_t cmd, uint16_t status, uint16_t length, uint8_t *data);