+++ CHANGELOG.md
@@ -4,6 +4,7 @@ This project uses the changelog in accordance with [keepchangelog](http://keepac
 
 ## [unreleased][unreleased]
 
+ - Added configurable wake-up time for button and field detection (1000-60000 ms) (@zeusricote)
  - Added commands to dump and clone Mifare tags
  - Improved slot LED indication when switching slots
  - Fixed hardnested subprocess call to work in read-only directories